@@ -1,6 +1,5 @@
 import {Injectable, NgZone} from '@angular/core';
-import {Observable} from 'rxjs/Observable';
-import {Observer} from 'rxjs/Observer';
+import {Observable, Observer} from 'rxjs';
 
 import * as mapTypes from './google-maps-types';
 import {Polyline} from './google-maps-types';
@@ -25,10 +24,12 @@ export class GoogleMapsAPIWrapper {
   }
 
   createMap(el: HTMLElement, mapOptions: mapTypes.MapOptions): Promise<void> {
-    return this._loader.load().then(() => {
-      const map = new google.maps.Map(el, mapOptions);
-      this._mapResolver(<mapTypes.GoogleMap>map);
-      return;
+    return this._zone.runOutsideAngular( () => {
+      return this._loader.load().then(() => {
+        const map = new google.maps.Map(el, mapOptions);
+        this._mapResolver(<mapTypes.GoogleMap>map);
+        return;
+      });
     });
   }
 
@@ -60,6 +61,16 @@ export class GoogleMapsAPIWrapper {
     return this._map.then((map: mapTypes.GoogleMap) => {
       options.map = map;
       return new google.maps.Circle(options);
+    });
+  }
+
+  /**
+   * Creates a google.map.Rectangle for the current map.
+   */
+  createRectangle(options: mapTypes.RectangleOptions): Promise<mapTypes.Rectangle> {
+    return this._map.then((map: mapTypes.GoogleMap) => {
+      options.map = map;
+      return new google.maps.Rectangle(options);
     });
   }
 
@@ -98,10 +109,16 @@ export class GoogleMapsAPIWrapper {
   }
 
   subscribeToMapEvent<E>(eventName: string): Observable<E> {
-    return Observable.create((observer: Observer<E>) => {
+    return new Observable((observer: Observer<E>) => {
       this._map.then((m: mapTypes.GoogleMap) => {
         m.addListener(eventName, (arg: E) => { this._zone.run(() => observer.next(arg)); });
       });
+    });
+  }
+
+  clearInstanceListeners() {
+    this._map.then((map: mapTypes.GoogleMap) => {
+      google.maps.event.clearInstanceListeners(map);
     });
   }
 
