@@ -1,14 +1,11 @@
-import {Directive, EventEmitter, OnChanges, OnDestroy, SimpleChange,
-  AfterContentInit, ContentChildren, QueryList, Input, Output
-} from '@angular/core';
-import {Subscription} from 'rxjs/Subscription';
-
-import {MouseEvent} from '../map-types';
+import { AfterContentInit, ContentChildren, Directive, EventEmitter, Input, OnChanges, OnDestroy, Output, QueryList, SimpleChange, forwardRef } from '@angular/core';
+import { Observable, ReplaySubject, Subscription } from 'rxjs';
+import { tap } from 'rxjs/operators';
+import { MarkerLabel, MouseEvent } from '../map-types';
+import { FitBoundsAccessor, FitBoundsDetails } from '../services/fit-bounds';
 import * as mapTypes from '../services/google-maps-types';
-import {MarkerManager} from '../services/managers/marker-manager';
-
-import {AgmInfoWindow} from './info-window';
-import {MarkerLabel} from '../map-types';
+import { MarkerManager } from '../services/managers/marker-manager';
+import { AgmInfoWindow } from './info-window';
 
 let markerId = 0;
 
@@ -37,13 +34,16 @@ let markerId = 0;
  */
 @Directive({
   selector: 'agm-marker',
+  providers: [
+    {provide: FitBoundsAccessor, useExisting: forwardRef(() => AgmMarker)}
+  ],
   inputs: [
     'latitude', 'longitude', 'title', 'label', 'draggable: markerDraggable', 'iconUrl',
     'openInfoWindow', 'opacity', 'visible', 'zIndex', 'animation'
   ],
   outputs: ['markerClick', 'dragEnd', 'mouseOver', 'mouseOut']
 })
-export class AgmMarker implements OnDestroy, OnChanges, AfterContentInit {
+export class AgmMarker implements OnDestroy, OnChanges, AfterContentInit, FitBoundsAccessor {
   /**
    * The latitude position of the marker.
    */
@@ -113,7 +113,12 @@ export class AgmMarker implements OnDestroy, OnChanges, AfterContentInit {
   /**
    * This event emitter gets emitted when the user clicks on the marker.
    */
-  @Output() markerClick: EventEmitter<void> = new EventEmitter<void>();
+  @Output() markerClick: EventEmitter<AgmMarker> = new EventEmitter<AgmMarker>();
+
+  /**
+   * This event is fired when the user rightclicks on the marker.
+   */
+  @Output() markerRightClick: EventEmitter<void> = new EventEmitter<void>();
 
   /**
    * This event is fired when the user stops dragging the marker.
@@ -139,6 +144,8 @@ export class AgmMarker implements OnDestroy, OnChanges, AfterContentInit {
   private _id: string;
   private _observableSubscriptions: Subscription[] = [];
 
+  protected readonly _fitBoundsDetails$: ReplaySubject<FitBoundsDetails> = new ReplaySubject<FitBoundsDetails>(1);
+
   constructor(private _markerManager: MarkerManager) { this._id = (markerId++).toString(); }
 
   /* @internal */
@@ -158,17 +165,25 @@ export class AgmMarker implements OnDestroy, OnChanges, AfterContentInit {
 
   /** @internal */
   ngOnChanges(changes: {[key: string]: SimpleChange}) {
+    if (typeof this.latitude === 'string') {
+      this.latitude = Number(this.latitude);
+    }
+    if (typeof this.longitude === 'string') {
+      this.longitude = Number(this.longitude);
+    }
     if (typeof this.latitude !== 'number' || typeof this.longitude !== 'number') {
       return;
     }
     if (!this._markerAddedToManger) {
       this._markerManager.addMarker(this);
+      this._updateFitBoundsDetails();
       this._markerAddedToManger = true;
       this._addEventListeners();
       return;
     }
     if (changes['latitude'] || changes['longitude']) {
       this._markerManager.updateMarkerPosition(this);
+      this._updateFitBoundsDetails();
     }
     if (changes['title']) {
       this._markerManager.updateTitle(this);
@@ -199,14 +214,30 @@ export class AgmMarker implements OnDestroy, OnChanges, AfterContentInit {
     }
   }
 
+  /**
+   * @internal
+   */
+  getFitBoundsDetails$(): Observable<FitBoundsDetails> {
+    return this._fitBoundsDetails$.asObservable();
+  }
+
+  protected _updateFitBoundsDetails() {
+    this._fitBoundsDetails$.next({latLng: {lat: this.latitude, lng: this.longitude}});
+  }
+
   private _addEventListeners() {
     const cs = this._markerManager.createEventObservable('click', this).subscribe(() => {
       if (this.openInfoWindow) {
         this.infoWindow.forEach(infoWindow => infoWindow.open());
       }
-      this.markerClick.emit(null);
+      this.markerClick.emit(this);
     });
     this._observableSubscriptions.push(cs);
+
+    const rc = this._markerManager.createEventObservable('rightclick', this).subscribe(() => {
+      this.markerRightClick.emit(null);
+    });
+    this._observableSubscriptions.push(rc);
 
     const ds =
         this._markerManager.createEventObservable<mapTypes.MouseEvent>('dragend', this)
