@@ -5,8 +5,8 @@ packages.forEach(pkgName => {
   console.log(`Copy assets for ${pkgName}`);
   const readmeFile = `./packages/${pkgName}/README.md`;
   if (fs.existsSync(readmeFile)) {
-    fs.writeFileSync(`dist/${pkgName}/README.md`, fs.readFileSync(readmeFile));
+    fs.writeFileSync(`dist/packages/${pkgName}/README.md`, fs.readFileSync(readmeFile));
   }
 
-  fs.writeFileSync(`dist/${pkgName}/LICENSE`, fs.readFileSync('LICENSE'));
+  fs.writeFileSync(`dist/packages/${pkgName}/LICENSE`, fs.readFileSync('LICENSE'));
 });
